+++ CHANGELOG.md
@@ -1,3 +1,4 @@
+- Add DataTable component [#137](https://github.com/PrefectHQ/miter-design/pull/137)
 - Add delete color option [#622](https://github.com/PrefectHQ/orion/issues/622)
 - Add Tags component [#697](https://github.com/PrefectHQ/orion/pull/697)
 - Fix bug where tabs do not update with programmatic setting of Vue ref - [#125](https://github.com/PrefectHQ/miter-design/pull/126)